# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos/me/{videoId}": {
      "get": {
        "description": "The owner status poll. Returns the row in any state, including `error` with its persisted failure reason, which is what makes a failed upload diagnosable by its owner instead of silently absent.",
        "operationId": "VideosController_findOwned",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "Internal video id returned by the initiate call",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The video, in whatever state it currently is",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "publicId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "draft",
                        "processing",
                        "ready",
                        "error"
                      ]
                    },
                    "failure_reason": {
                      "type": "string",
                      "nullable": true,
                      "description": "Populated when `status` is `error`"
                    },
                    "duration_seconds": {
                      "type": "number",
                      "nullable": true,
                      "example": 12.345
                    },
                    "width": {
                      "type": "integer",
                      "nullable": true,
                      "example": 1920
                    },
                    "height": {
                      "type": "integer",
                      "nullable": true,
                      "example": 1080
                    },
                    "video_codec": {
                      "type": "string",
                      "nullable": true,
                      "example": "h264"
                    },
                    "audio_codec": {
                      "type": "string",
                      "nullable": true,
                      "example": "aac"
                    },
                    "container_format": {
                      "type": "string",
                      "nullable": true,
                      "example": "mp4"
                    },
                    "bitrate_bps": {
                      "type": "integer",
                      "nullable": true,
                      "example": 4500000
                    },
                    "size_bytes": {
                      "type": "integer",
                      "nullable": true,
                      "example": 47225
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND — unknown video or a caller who does not own it; the two are deliberately indistinguishable, and a 403 would confirm existence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Read one of the caller's own videos, in any state",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}": {
      "get": {
        "description": "Anonymous route. Resolves the public identifier filtering on `status = ready` in the same query, so a video that is not ready is indistinguishable from one that does not exist.",
        "operationId": "VideosController_findPublic",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Short public identifier of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metadata of a ready video; `status` is ready by construction",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "publicId": {
                      "type": "string"
                    },
                    "duration_seconds": {
                      "type": "number",
                      "nullable": true,
                      "example": 12.345
                    },
                    "width": {
                      "type": "integer",
                      "nullable": true,
                      "example": 1920
                    },
                    "height": {
                      "type": "integer",
                      "nullable": true,
                      "example": 1080
                    },
                    "video_codec": {
                      "type": "string",
                      "nullable": true,
                      "example": "h264"
                    },
                    "audio_codec": {
                      "type": "string",
                      "nullable": true,
                      "example": "aac"
                    },
                    "container_format": {
                      "type": "string",
                      "nullable": true,
                      "example": "mp4"
                    },
                    "bitrate_bps": {
                      "type": "integer",
                      "nullable": true,
                      "example": 4500000
                    },
                    "size_bytes": {
                      "type": "integer",
                      "nullable": true,
                      "example": 47225
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND — unknown identifier or a video that is not ready; the two answer identically so the route is no existence oracle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Read the public metadata of a ready video",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/stream": {
      "get": {
        "description": "The API never carries the video bytes: it answers `302` and the player talks to the storage server directly, which is also what gets `Range`/`206` semantics for free instead of hand-rolling partial content.",
        "operationId": "VideosController_stream",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Short public identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Presigned URL in `Location`, valid for minutes. The response body is empty — the bytes come from the storage server.",
            "headers": {
              "Location": {
                "description": "Short-lived presigned `GET` URL",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "Cache-Control": {
                "description": "Caches the redirect, never the presigned object — the signature rotates per request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND — unknown identifier or a video that is not ready; the three delivery routes inherit the metadata route’s `ready`-only filter verbatim, so none of them is an existence oracle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Redirect to a short-lived presigned URL for playback",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/download": {
      "get": {
        "description": "Same object as `/stream`; the difference is the `response-content-disposition` carried into the signature.",
        "operationId": "VideosController_download",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Short public identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Presigned URL in `Location`, valid for minutes. The response body is empty — the bytes come from the storage server.",
            "headers": {
              "Location": {
                "description": "Short-lived presigned `GET` URL",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "Cache-Control": {
                "description": "Caches the redirect, never the presigned object — the signature rotates per request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND — unknown identifier or a video that is not ready; the three delivery routes inherit the metadata route’s `ready`-only filter verbatim, so none of them is an existence oracle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Redirect to the same object, signed as an attachment",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/thumbnail": {
      "get": {
        "description": "The content type is fixed at signing time, so the browser renders the image inline whatever the worker wrote on the object. The `Cache-Control` sits on this `302`, never on the image: the signature rotates per request, so the image itself can never be cached.",
        "operationId": "VideosController_thumbnail",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Short public identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Presigned URL in `Location`, valid for minutes. The response body is empty — the bytes come from the storage server.",
            "headers": {
              "Location": {
                "description": "Short-lived presigned `GET` URL",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "Cache-Control": {
                "description": "Caches the redirect, never the presigned object — the signature rotates per request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND — unknown identifier or a video that is not ready; the three delivery routes inherit the metadata route’s `ready`-only filter verbatim, so none of them is an existence oracle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Redirect to a presigned thumbnail URL pinned to `image/jpeg`",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}